% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlson.R
\name{charlson}
\alias{charlson}
\title{Charlson score table}
\usage{
apache.ii(data, id = names(data)[1],
which = c('myocardial_infarct', 'congestive_heart_failure', 'peripheral_vascular_disease',
cerebrovascular_disease', 'dementia', 'chronic_pulmonary_disease',
connective_tissue_disease', 'ulcer_disease', 'mild_liver_disease', 'diabetes',
hemiplegia', 'moderate_or_severe_renal_disease', 'diabetes_with_end_organ_damage', 'any_tumor',
moderate_or_severe_liver_disease', 'metastatic_solid_tumor', 'AIDS'),
...)
}
\arguments{
\item{data}{An object of class data.frame.
This should have either:

(1) Matched names with the confounders listed in which. Each would serve as one confounder in the score table or

(2) Matched orders with the confounders listed in which.}

\item{id}{Column name of column id in data serving as subject id. Default is the first one.
Set to NULL if the data doesn't have ID column.}

\item{which}{Which confounders will be calculated.}
}
\value{
A tibble that contains scores for each confounders and total score.
}
\description{
This function calculation the Charlson table
}
\seealso{
\link{summary.score_tbl}.
}
