% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTable.R
\name{flextable.ScoreTable}
\alias{flextable.ScoreTable}
\title{A method to convert score table to flextable}
\usage{
\method{flextable}{ScoreTable}(x, ...)
}
\arguments{
\item{x}{An object of class ScoreTable}

\item{...}{Additional function passed to flextable::flextable.}
}
\description{
A method to convert score table to flextable
}
\seealso{
\link{print.ScoreTable}, \link{huxtable.ScoreTable}, \link{as.data.frame.ScoreTable},
 \link[flextable]{flextable}
}
