% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTable.R
\name{huxtable.ScoreTable}
\alias{huxtable.ScoreTable}
\title{A method to convert score table to huxtable}
\usage{
\method{huxtable}{ScoreTable}(x, ...)
}
\arguments{
\item{x}{An object of class ScoreTable}

\item{...}{Additional function passed to huxtable::hux.}
}
\description{
A method to convert score table to huxtable
}
\seealso{
\link{print.ScoreTable}, \link{huxtable.ScoreTable}, \link{as.data.frame.ScoreTable},
\link[huxtable]{huxtable}
}
