% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTable.R
\name{print.ScoreTable}
\alias{print.ScoreTable}
\title{A method to print out ScoreTable object}
\usage{
\method{print}{ScoreTable}(x, pretty = TRUE, ...)
}
\arguments{
\item{x}{An object of class ScoreTable}

\item{pretty}{A logical value. Default = TRUE will print out the pretty version of the table.}
}
\description{
A method to print out ScoreTable object
}
\seealso{
\link{as.data.frame.ScoreTable}, \link{huxtable.ScoreTable}, \link{flextable.ScoreTable}
}
