% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{summary.score_tbl}
\alias{summary.score_tbl}
\title{Summary a score_tbl object}
\usage{
\method{summary}{score_tbl}(x, method = NULL, ...)
}
\arguments{
\item{x}{An object of class ScoreTable}

\item{method}{A function to summarise the table.
For example: max, mean, min.
The function must return unique value for each ID.
This will be passed as .funs in \link[dplyr]{summarise_all}}

\item{...}{Addition params passed to method}
}
\value{
A tibble
}
\description{
A function to do summarisation on a calculated score table.
}
