% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache.ii.R
\name{apache.ii}
\alias{apache.ii}
\title{APACHE II score table}
\usage{
apache.ii(data, id = names(data)[1],
which = c('temp', 'map', 'hr', 'rr', 'aapo2', 'pao2', 'ph', 'hco3', 'sodium', 'potassium', 'creatinine', 'hct', 'wbc',
'gcs', 'age', 'chronic'),
...)
}
\arguments{
\item{data}{An object of class data.frame.
This should have either:

(1) Matched names with the confounders listed in which. Each would serve as one confounder in the score table or

(2) Matched orders with the confounders listed in which.}

\item{id}{Column name of column id in data serving as subject id. Default is the first one.
Set to NULL if the data doesn't have ID column.}

\item{which}{Which confounders will be calculated.}
}
\value{
A tibble that contains scores for each confounders and total score.
}
\description{
This function calculates the APACHE II score table
}
\references{
Knaus WA, Draper EA, Wagner DP, Zimmerman JE.
APACHE II: a severity of disease classification system. Crit Care Med. 1985 Oct;13(10):818-29.
PMID: \href{https://www.ncbi.nlm.nih.gov/pubmed/3928249}{3928249}
}
\seealso{
\link{summary.score_tbl}.
}
