% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTable.R
\name{ScoreTable}
\alias{ScoreTable}
\title{Score Table generator}
\usage{
ScoreTable(confounders, scores, aliases = NULL, custom_cases)
}
\arguments{
\item{confounders}{For simple generation.
A character vector that declares derived components in the score table (1) or
a named list (2) whose names define the derived components in the score table and
whose values define their respective possible levels}

\item{scores}{For simple generation.
A numeric vector that declares score for derived components using method (1) or
a named list whose names follow the derived component names defined in method (2) and
whose values define their respective possible scores}

\item{aliases}{A named list that define a pretty representatives for defined confounders
following the structure of list(confounder = aliases).
Unmentioned confounders will be left intact.}

\item{custom_cases}{A named list that define a sophisticated way to define confounders and scoring algorithm,
following the structure of list(name = list(formulas))
Each name is the name for derived confounders.
Each formula in each sub-list folllows the form of condition ~ score in a "specific to general" order.
This is based on \link[dplyr]{case_when}.}
}
\value{
An object of class ScoreTable.
When called with no data, this will print out the structure of the ScoreTable.
When called with data passed, this will return a data frame of class score_tbl.
}
\description{
A function to formally create an object of class Score Table.
}
\seealso{
\link[dplyr]{case_when}, \link{as.data.frame.ScoreTable}, \link{apache.ii}, \link{summary.score_tbl}
}
